chore: ajusta arquivo do airflow
