+++ cyber-gen/.env
@@ -1,9 +1,16 @@
-DB_URL="postgresql://postgres:postgres@localhost:5432/postgres"
-COUCHDB_URL="http://admin:password@localhost:5984"
-COUCHDB_DBNAME="etl_jobs"
+# PostgreSQL Database
+DB_URL="postgresql://tiagornandrade:qTGNI8aumD5Z@ep-rapid-scene-63978506-pooler.us-east-2.aws.neon.tech/biaas?sslmode=require"
+
+# Firestore (substituindo CouchDB)
+GCP_PROJECT_ID="yams-lab-nonprod"  # Substitua pelo ID do projeto no GCP
+FIRESTORE_COLLECTION="etl_jobs"
+
+# Configuração da Cloud Function ou Cloud Run
 PROJECT_NAME="etl_loader"
 FUNCTION_NAME="etl_loader"
 GCP_REGION="us-central1"
 PORT="5555"
+
+# Configuração do Airflow (se aplicável)
 AIRFLOW_UID=50000
 AIRFLOW_GID=0
