+++ cyber-gen/.gitignore
@@ -1,0 +1,3 @@
+credential/
+.env
+/.env
