+++ cyber-gen/src/function_api.py
@@ -1,14 +1,16 @@
+import os
 from fastapi import FastAPI, HTTPException
 from datetime import datetime
+import uvicorn
 from src.utils.query_manager import execute_query
-from src.utils.couchdb import create_status_record, update_status, db
+from src.utils.firestore import create_status_record, update_status, db
 
 app = FastAPI()
 
 
 def process_job(layer: str, table: str):
     """
-    Processa a execução da query da camada específica e atualiza o status no CouchDB.
+    Processa a execução da query da camada específica e atualiza o status no Datastore.
     """
     doc_id = None
 
@@ -35,6 +37,11 @@ def process_job(layer: str, table: str):
         raise HTTPException(status_code=500, detail=str(e))
 
 
+@app.get("/")
+def read_root():
+    return {"message": "Cyber-gen is running"}
+
+
 @app.post("/{layer}/{table}/execute")
 def execute_table(layer: str, table: str):
     """Executa a query correspondente à camada e registra o status."""
@@ -48,8 +55,17 @@ def execute_table(layer: str, table: str):
 def get_status(job_id: str):
     """Retorna o status de um job pelo ID."""
     try:
-        if job_id in db:
-            return db[job_id]
+        task_key = db.key("etl_jobs", job_id)
+        entity = db.get(task_key)
+
+        if entity:
+            return entity
         raise HTTPException(status_code=404, detail="Job não encontrado")
+
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
+
+
+if __name__ == "__main__":
+    port = int(os.getenv("PORT", 8080))
+    uvicorn.run(app, host="0.0.0.0", port=port)
