+++ hiper-flow/logs/dag_id=load_bronze_accounts_deleted/run_id=scheduled__2025-02-25T00:00:00+00:00/task_id=bronze_accounts_deleted_trigger/attempt=20.log
@@ -1,0 +1,16 @@
+[2025-02-27T20:07:52.318+0000] {taskinstance.py:1957} INFO - Dependencies all met for dep_context=non-requeueable deps ti=<TaskInstance: load_bronze_accounts_deleted.bronze_accounts_deleted_trigger scheduled__2025-02-25T00:00:00+00:00 [queued]>
+[2025-02-27T20:07:52.321+0000] {taskinstance.py:1957} INFO - Dependencies all met for dep_context=requeueable deps ti=<TaskInstance: load_bronze_accounts_deleted.bronze_accounts_deleted_trigger scheduled__2025-02-25T00:00:00+00:00 [queued]>
+[2025-02-27T20:07:52.322+0000] {taskinstance.py:2171} INFO - Starting attempt 20 of 21
+[2025-02-27T20:07:52.329+0000] {taskinstance.py:2192} INFO - Executing <Task(SimpleHttpOperator): bronze_accounts_deleted_trigger> on 2025-02-25 00:00:00+00:00
+[2025-02-27T20:07:52.331+0000] {standard_task_runner.py:60} INFO - Started process 153 to run task
+[2025-02-27T20:07:52.333+0000] {standard_task_runner.py:87} INFO - Running: ['***', 'tasks', 'run', 'load_bronze_accounts_deleted', 'bronze_accounts_deleted_trigger', 'scheduled__2025-02-25T00:00:00+00:00', '--job-id', '62', '--raw', '--subdir', 'DAGS_FOLDER/dynamic_api_trigger.py', '--cfg-path', '/tmp/tmptplrz052']
+[2025-02-27T20:07:52.334+0000] {standard_task_runner.py:88} INFO - Job 62: Subtask bronze_accounts_deleted_trigger
+[2025-02-27T20:07:52.357+0000] {task_command.py:423} INFO - Running <TaskInstance: load_bronze_accounts_deleted.bronze_accounts_deleted_trigger scheduled__2025-02-25T00:00:00+00:00 [running]> on host e37045c436ab
+[2025-02-27T20:07:52.391+0000] {taskinstance.py:2481} INFO - Exporting env vars: AIRFLOW_CTX_DAG_OWNER='***' AIRFLOW_CTX_DAG_ID='load_bronze_accounts_deleted' AIRFLOW_CTX_TASK_ID='bronze_accounts_deleted_trigger' AIRFLOW_CTX_EXECUTION_DATE='2025-02-25T00:00:00+00:00' AIRFLOW_CTX_TRY_NUMBER='20' AIRFLOW_CTX_DAG_RUN_ID='scheduled__2025-02-25T00:00:00+00:00'
+[2025-02-27T20:07:52.392+0000] {http.py:169} INFO - Calling HTTP method
+[2025-02-27T20:07:52.395+0000] {base.py:83} INFO - Using connection ID 'api-etl' for task execution.
+[2025-02-27T20:07:52.398+0000] {base.py:83} INFO - Using connection ID 'api-etl' for task execution.
+[2025-02-27T20:07:53.249+0000] {http.py:208} INFO - {"status":"Execução concluída","job_id":"accounts_deleted_execute_20250227T200752"}
+[2025-02-27T20:07:53.272+0000] {taskinstance.py:1138} INFO - Marking task as SUCCESS. dag_id=load_bronze_accounts_deleted, task_id=bronze_accounts_deleted_trigger, execution_date=20250225T000000, start_date=20250227T200752, end_date=20250227T200753
+[2025-02-27T20:07:53.323+0000] {local_task_job_runner.py:234} INFO - Task exited with return code 0
+[2025-02-27T20:07:53.342+0000] {taskinstance.py:3281} INFO - 0 downstream tasks scheduled from follow-on schedule check
