+++ hiper-flow/logs/dag_id=load_bronze_accounts_deleted/run_id=scheduled__2025-02-26T00:00:00+00:00/task_id=bronze_accounts_deleted_trigger/attempt=2.log
@@ -1,0 +1,29 @@
+[2025-02-27T20:05:09.411+0000] {taskinstance.py:1957} INFO - Dependencies all met for dep_context=non-requeueable deps ti=<TaskInstance: load_bronze_accounts_deleted.bronze_accounts_deleted_trigger scheduled__2025-02-26T00:00:00+00:00 [queued]>
+[2025-02-27T20:05:09.414+0000] {taskinstance.py:1957} INFO - Dependencies all met for dep_context=requeueable deps ti=<TaskInstance: load_bronze_accounts_deleted.bronze_accounts_deleted_trigger scheduled__2025-02-26T00:00:00+00:00 [queued]>
+[2025-02-27T20:05:09.414+0000] {taskinstance.py:2171} INFO - Starting attempt 2 of 3
+[2025-02-27T20:05:09.420+0000] {taskinstance.py:2192} INFO - Executing <Task(SimpleHttpOperator): bronze_accounts_deleted_trigger> on 2025-02-26 00:00:00+00:00
+[2025-02-27T20:05:09.422+0000] {standard_task_runner.py:60} INFO - Started process 101 to run task
+[2025-02-27T20:05:09.424+0000] {standard_task_runner.py:87} INFO - Running: ['***', 'tasks', 'run', 'load_bronze_accounts_deleted', 'bronze_accounts_deleted_trigger', 'scheduled__2025-02-26T00:00:00+00:00', '--job-id', '60', '--raw', '--subdir', 'DAGS_FOLDER/dynamic_api_trigger.py', '--cfg-path', '/tmp/tmpxhjjm87t']
+[2025-02-27T20:05:09.425+0000] {standard_task_runner.py:88} INFO - Job 60: Subtask bronze_accounts_deleted_trigger
+[2025-02-27T20:05:09.447+0000] {task_command.py:423} INFO - Running <TaskInstance: load_bronze_accounts_deleted.bronze_accounts_deleted_trigger scheduled__2025-02-26T00:00:00+00:00 [running]> on host e37045c436ab
+[2025-02-27T20:05:09.483+0000] {taskinstance.py:2481} INFO - Exporting env vars: AIRFLOW_CTX_DAG_OWNER='***' AIRFLOW_CTX_DAG_ID='load_bronze_accounts_deleted' AIRFLOW_CTX_TASK_ID='bronze_accounts_deleted_trigger' AIRFLOW_CTX_EXECUTION_DATE='2025-02-26T00:00:00+00:00' AIRFLOW_CTX_TRY_NUMBER='2' AIRFLOW_CTX_DAG_RUN_ID='scheduled__2025-02-26T00:00:00+00:00'
+[2025-02-27T20:05:09.483+0000] {http.py:169} INFO - Calling HTTP method
+[2025-02-27T20:05:09.490+0000] {taskinstance.py:2699} ERROR - Task failed with exception
+Traceback (most recent call last):
+  File "/home/airflow/.local/lib/python3.8/site-packages/airflow/models/taskinstance.py", line 433, in _execute_task
+    result = execute_callable(context=context, **execute_callable_kwargs)
+  File "/home/airflow/.local/lib/python3.8/site-packages/airflow/providers/http/operators/http.py", line 166, in execute
+    return self.execute_sync(context=context)
+  File "/home/airflow/.local/lib/python3.8/site-packages/airflow/providers/http/operators/http.py", line 170, in execute_sync
+    response = self.hook.run(self.endpoint, self.data, self.headers, self.extra_options)
+  File "/home/airflow/.local/lib/python3.8/site-packages/airflow/providers/http/operators/http.py", line 148, in hook
+    conn = BaseHook.get_connection(conn_id)
+  File "/home/airflow/.local/lib/python3.8/site-packages/airflow/hooks/base.py", line 82, in get_connection
+    conn = Connection.get_connection_from_secrets(conn_id)
+  File "/home/airflow/.local/lib/python3.8/site-packages/airflow/models/connection.py", line 479, in get_connection_from_secrets
+    raise AirflowNotFoundException(f"The conn_id `{conn_id}` isn't defined")
+airflow.exceptions.AirflowNotFoundException: The conn_id `api-etl` isn't defined
+[2025-02-27T20:05:09.492+0000] {taskinstance.py:1138} INFO - Marking task as UP_FOR_RETRY. dag_id=load_bronze_accounts_deleted, task_id=bronze_accounts_deleted_trigger, execution_date=20250226T000000, start_date=20250227T200509, end_date=20250227T200509
+[2025-02-27T20:05:09.499+0000] {standard_task_runner.py:107} ERROR - Failed to execute job 60 for task bronze_accounts_deleted_trigger (The conn_id `api-etl` isn't defined; 101)
+[2025-02-27T20:05:09.520+0000] {local_task_job_runner.py:234} INFO - Task exited with return code 1
+[2025-02-27T20:05:09.528+0000] {taskinstance.py:3281} INFO - 0 downstream tasks scheduled from follow-on schedule check
