+++ hiper-flow/logs/dag_id=load_bronze_accounts_deleted/run_id=scheduled__2025-02-26T00:00:00+00:00/task_id=bronze_accounts_deleted_trigger/attempt=3.log
@@ -1,0 +1,16 @@
+[2025-02-27T20:06:21.341+0000] {taskinstance.py:1957} INFO - Dependencies all met for dep_context=non-requeueable deps ti=<TaskInstance: load_bronze_accounts_deleted.bronze_accounts_deleted_trigger scheduled__2025-02-26T00:00:00+00:00 [queued]>
+[2025-02-27T20:06:21.345+0000] {taskinstance.py:1957} INFO - Dependencies all met for dep_context=requeueable deps ti=<TaskInstance: load_bronze_accounts_deleted.bronze_accounts_deleted_trigger scheduled__2025-02-26T00:00:00+00:00 [queued]>
+[2025-02-27T20:06:21.345+0000] {taskinstance.py:2171} INFO - Starting attempt 3 of 4
+[2025-02-27T20:06:21.352+0000] {taskinstance.py:2192} INFO - Executing <Task(SimpleHttpOperator): bronze_accounts_deleted_trigger> on 2025-02-26 00:00:00+00:00
+[2025-02-27T20:06:21.354+0000] {standard_task_runner.py:60} INFO - Started process 127 to run task
+[2025-02-27T20:06:21.356+0000] {standard_task_runner.py:87} INFO - Running: ['***', 'tasks', 'run', 'load_bronze_accounts_deleted', 'bronze_accounts_deleted_trigger', 'scheduled__2025-02-26T00:00:00+00:00', '--job-id', '61', '--raw', '--subdir', 'DAGS_FOLDER/dynamic_api_trigger.py', '--cfg-path', '/tmp/tmp8pxca8dv']
+[2025-02-27T20:06:21.357+0000] {standard_task_runner.py:88} INFO - Job 61: Subtask bronze_accounts_deleted_trigger
+[2025-02-27T20:06:21.379+0000] {task_command.py:423} INFO - Running <TaskInstance: load_bronze_accounts_deleted.bronze_accounts_deleted_trigger scheduled__2025-02-26T00:00:00+00:00 [running]> on host e37045c436ab
+[2025-02-27T20:06:21.419+0000] {taskinstance.py:2481} INFO - Exporting env vars: AIRFLOW_CTX_DAG_OWNER='***' AIRFLOW_CTX_DAG_ID='load_bronze_accounts_deleted' AIRFLOW_CTX_TASK_ID='bronze_accounts_deleted_trigger' AIRFLOW_CTX_EXECUTION_DATE='2025-02-26T00:00:00+00:00' AIRFLOW_CTX_TRY_NUMBER='3' AIRFLOW_CTX_DAG_RUN_ID='scheduled__2025-02-26T00:00:00+00:00'
+[2025-02-27T20:06:21.419+0000] {http.py:169} INFO - Calling HTTP method
+[2025-02-27T20:06:21.423+0000] {base.py:83} INFO - Using connection ID 'api-etl' for task execution.
+[2025-02-27T20:06:21.426+0000] {base.py:83} INFO - Using connection ID 'api-etl' for task execution.
+[2025-02-27T20:06:22.316+0000] {http.py:208} INFO - {"status":"Execução concluída","job_id":"accounts_deleted_execute_20250227T200622"}
+[2025-02-27T20:06:22.330+0000] {taskinstance.py:1138} INFO - Marking task as SUCCESS. dag_id=load_bronze_accounts_deleted, task_id=bronze_accounts_deleted_trigger, execution_date=20250226T000000, start_date=20250227T200621, end_date=20250227T200622
+[2025-02-27T20:06:22.369+0000] {local_task_job_runner.py:234} INFO - Task exited with return code 0
+[2025-02-27T20:06:22.378+0000] {taskinstance.py:3281} INFO - 0 downstream tasks scheduled from follow-on schedule check
