+++ hiper-flow/logs/scheduler/latest
@@ -1,1 +1,1 @@
-/opt/airflow/logs/scheduler/2025-02-26
+/opt/airflow/logs/scheduler/2025-02-27
